requirements.txt
python-dotenv~=1.0.1
selenium~=4.17.2
djangorestframework~=3.14.0
webdriver-manager
python-dotenv
requests~=2.31.0
Django~=5.0.1
python-json-logger
pygsheets